+++ plugins/tflite/tflite.rs
@@ -42,11 +42,7 @@ use std::{
     time::Duration,
 };
 use thiserror::Error;
-use tokio::{
-    io::AsyncWriteExt,
-    runtime::Handle,
-    sync::{mpsc, Mutex},
-};
+use tokio::{io::AsyncWriteExt, runtime::Handle, sync::mpsc};
 use tokio_util::sync::CancellationToken;
 use url::Url;
 
@@ -96,7 +92,7 @@ pub struct TflitePlugin {
     _shutdown_complete_tx: mpsc::Sender<()>,
     logger: DynLogger,
     auth: DynAuth,
-    monitor_manager: Arc<Mutex<MonitorManager>>,
+    monitor_manager: MonitorManager,
     detector_manager: DetectorManager,
 }
 
@@ -107,7 +103,7 @@ impl TflitePlugin {
         logger: DynLogger,
         env: DynEnvConfig,
         auth: DynAuth,
-        monitor_manager: Arc<Mutex<MonitorManager>>,
+        monitor_manager: MonitorManager,
     ) -> Self {
         let tflite_logger = Arc::new(TfliteLogger {
             logger: logger.clone(),
@@ -175,7 +171,6 @@ impl Plugin for TflitePlugin {
 
     fn route(&self, router: Router) -> Router {
         let state = HandlerState {
-            rt_handler: self.rt_handle.clone(),
             logger: self.logger.clone(),
             monitor_manager: self.monitor_manager.clone(),
         };
@@ -786,18 +781,15 @@ impl Fetcher for Fetch {
 
 #[derive(Clone)]
 struct HandlerState {
-    rt_handler: Handle,
     logger: DynLogger,
-    monitor_manager: Arc<Mutex<MonitorManager>>,
+    monitor_manager: MonitorManager,
 }
 
 async fn enable_handler(
     State(s): State<HandlerState>,
     Path(monitor_id): Path<MonitorId>,
 ) -> Response {
-    let mut manager = s.monitor_manager.lock().await;
-
-    let Some(old_config) = manager.monitor_config(&monitor_id) else {
+    let Some(old_config) = s.monitor_manager.monitor_config(monitor_id.clone()).await else {
         return (
             StatusCode::NOT_FOUND,
             format!("monitor '{monitor_id}' does not exist"),
@@ -805,7 +797,7 @@ async fn enable_handler(
             .into_response();
     };
 
-    let Some(new_config) = set_enable(old_config, true) else {
+    let Some(new_config) = set_enable(&old_config, true) else {
         return (
             StatusCode::INTERNAL_SERVER_ERROR,
             "failed to find enable field",
@@ -813,11 +805,7 @@ async fn enable_handler(
             .into_response();
     };
 
-    match manager.monitor_set(&s.rt_handler, new_config).await {
-        Ok(created) => assert!(!created),
-        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response(),
-    }
-    if let Err(e) = manager.monitor_restart(&monitor_id).await {
+    if let Err(e) = s.monitor_manager.monitor_set_and_restart(new_config).await {
         return (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response();
     }
 
@@ -835,9 +823,7 @@ async fn disable_handler(
     State(s): State<HandlerState>,
     Path(monitor_id): Path<MonitorId>,
 ) -> Response {
-    let mut manager = s.monitor_manager.lock().await;
-
-    let Some(old_config) = manager.monitor_config(&monitor_id) else {
+    let Some(old_config) = s.monitor_manager.monitor_config(monitor_id.clone()).await else {
         return (
             StatusCode::NOT_FOUND,
             format!("monitor '{monitor_id}' does not exist"),
@@ -845,17 +831,13 @@ async fn disable_handler(
             .into_response();
     };
 
-    let new_config = match set_enable(old_config, false) {
+    let new_config = match set_enable(&old_config, false) {
         Some(v) => v,
         // None means that it's already disabled in some way.
         None => old_config.clone(),
     };
 
-    match manager.monitor_set(&s.rt_handler, new_config).await {
-        Ok(created) => assert!(!created),
-        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response(),
-    }
-    if let Err(e) = manager.monitor_restart(&monitor_id).await {
+    if let Err(e) = s.monitor_manager.monitor_set_and_restart(new_config).await {
         return (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response();
     }