+++ src/plugin/plugin.rs
@@ -17,10 +17,7 @@ use std::{
     sync::Arc,
 };
 use thiserror::Error;
-use tokio::{
-    runtime::Handle,
-    sync::{mpsc, Mutex},
-};
+use tokio::{runtime::Handle, sync::mpsc};
 use tokio_util::sync::CancellationToken;
 use types::{Assets, NewAuthFn, Templates};
 
@@ -58,7 +55,7 @@ pub trait Plugin {
 pub trait Application {
     fn rt_handle(&self) -> Handle;
     fn auth(&self) -> DynAuth;
-    fn monitor_manager(&self) -> Arc<Mutex<MonitorManager>>;
+    fn monitor_manager(&self) -> MonitorManager;
     fn shutdown_complete_tx(&self) -> mpsc::Sender<()>;
     fn logger(&self) -> DynLogger;
     fn env(&self) -> DynEnvConfig;